+++ packages/kernel/source/ipc/index.ts
@@ -1,3 +1,3 @@
 export * from "./emit.js";
 export * from "./handler.js";
-export * from "./sub-process.js";
+export * from "./subprocess.js";
